fix 104 test failures: 28 real port bugs + stale test expectations
Multi-agent triage of the failures the restored suite surfaced. Real production bugs fixed
(examples): sarif reference types cast enum value() as u8 turning -1 into 255; profile_constants
to_string used Rust trim() which unlike Java doesn't strip NUL/control chars; SleighLanguage::decode
built the address factory with wrong default space; RefCell double-borrows; ext4 provider fallthrough.
Rest were stale test expectations aligned to current Java-derived behavior. Full-suite verify next.

+++ ghidra-rs/src/sarif/export/ref/ext_reference.rs
@@ -19,7 +19,7 @@ impl ExtReference {
     /// from the given reference and stores them as strings.
     pub fn new(reference: &dyn Reference) -> Self {
         let reference_type = reference.reference_type();
-        let index = (reference_type.value() as u8).to_string();
+        let index = reference_type.value().to_string();
         let kind = reference_type.name().to_string();
         let op_index = reference.operand_index();
         let source_type = reference.source().display_string().to_string();
