+++ ghidra-rs/src/sarif/export/ref/ext_stack_reference.rs
@@ -21,7 +21,7 @@ impl ExtStackReference {
     /// from the given stack reference.
     pub fn new(reference: &dyn StackReference) -> Self {
         let reference_type = reference.reference_type();
-        let index = (reference_type.value() as u8).to_string();
+        let index = reference_type.value().to_string();
         let kind = reference_type.name().to_string();
         let op_index = reference.operand_index();
         let source_type = reference.source().display_string().to_string();
